Tighten trace support review fixes
Gate recorded-trace replay on `libsql`, move trace-template tests out of
the helper module, and add context to trace fixture loading errors.

Also document support module exports and clean up the developer guide table
caption so the Markdown and Rust gates stay aligned with the rebased branch.
