+++ tests/e2e_traces/recorded_trace.rs
@@ -3,15 +3,18 @@
 //! Each test replays a recorded fixture through the full agent loop, verifying
 //! declarative `expects` from the JSON and any additional manual checks.
 
+#[cfg(feature = "libsql")]
 use crate::support::run_recorded_trace;
+#[cfg(feature = "libsql")]
 use rstest::rstest;
 
 /// Recorded trace tests covering telegram check, weather query, and baseball stats.
+#[cfg(feature = "libsql")]
 #[rstest]
 #[case("telegram_check.json")]
 #[case("weather_sf.json")]
 #[case("baseball_stats.json")]
 #[tokio::test]
-async fn recorded_trace(#[case] fixture_name: &str) {
-    run_recorded_trace(fixture_name).await;
+async fn recorded_trace(#[case] fixture_name: &str) -> anyhow::Result<()> {
+    run_recorded_trace(fixture_name).await
 }
